pyproject.toml
[build-system]
requires = ["pdm-pep517>=1.0.0", "numpy>=1.23.5"]
build-backend = "pdm.pep517.api"

[project]
name = "cyto-dl"
version = "0.4.0"
description = """\
  Collection of representation learning models, techniques, callbacks, utils, \
  used to create latent variable models of cell shape, morphology and \
  intracellular organization.\
  """
readme = "README.md"
authors = [
    { name = "Benji Morris", email = "benjamin.morris@alleninstitute.org" },
    { name = "Guilherme Pires", email = "guilherme.pires@alleninstitute.org" },
    { name = "Ritvik Vasan", email = "ritvik.vasan@alleninstitute.org" },
]
dependencies = [
    "hydra-core~=1.3.0",
    "hydra-colorlog>=1.2",
    "hydra-optuna-sweeper>=1.2",
    "torch~=2.0.0",
    "numpy>=1.23",
    "matplotlib>=3.7",
    "pandas>=1.5",
    "fire>=0.5",
    "mlflow>=2.1",
    "omegaconf>=2.3",
    "pyarrow>=10.0",
    "pyrootutils>=1.0",
    "PyYAML>=6.0",
    "scikit-learn>=1.2",
    "universal-pathlib>=0.0",
    "ome-zarr>=0.6",
    "anndata>=0.8",
    "monai-weekly>=1.2.dev2308",
    "timm>=0.9.7",
    "tqdm>=4.64",
    "lightning>=2.0",
    "einops>=0.6.1",
    "edt>=2.3.1",
    "astropy>=5.2",
    "rich",
    "boto3",
    "bioio>=1.0.1",
    "bioio-czi",
    "bioio-ome-tiff",
    "bioio-tifffile",
    "online-stats>=2023",
    "positional-encodings>=6.0.3",
]
requires-python = ">=3.9,<3.11"

[project.optional-dependencies]
equiv = [
    "lie_learn==0.0.1.post1",
    "escnn~=1.0.7",
    "py3nj==0.1.2",
    "e3nn~=0.5.1"
]
spharm = [
    "vtk~=9.2",
    "aicscytoparam~=0.1",
    "pyshtools==4.10.3",
]
s3 = [
    "boto3>=1.23.5,<1.24.5",
    "s3fs~=2023.1"
]
torchserve = [
    "torchserve>=0.8.0",
    "gorilla>=0.4.0",
]
pcloud = [
    "pyntcloud>=0.3.1",
    "aicsshparam>=0.1.7",
    "torch-geometric>=2.3.1",
    "point-cloud-utils>=0.29.6",
    "geomloss>=0.2.6",
    "Ninja>=1.11.1",
    "torchio>=0.19.1",
]
all = [
    "cyto-dl[equiv,spharm,s3,torchserve,pcloud]",
]
test = [
    "cyto-dl",
    "pytest~=7.2",
    "pytest-cov[toml]~=4.0",
]
docs = [
    "cyto-dl",
    "furo<=2023.3.23",
    "m2r2<=0.3.3.post2",
    "sphinx<=6.1.3",
]


[project.urls]
Homepage = "https://github.com/AllenCellModeling/cyto-dl"

[tool.pdm.build]
includes = ["cyto_dl", "README.md", "LICENSE", "**/*.yaml" ]
excludes = ["**/*.pyc", "**/__pycache__"]

[tool.flake8]
ignore = [
    "E203",
    "E402",
    "W291",
    "W503",
]
max-line-length = 88

[tool.pytest.ini_options]
addopts = [
    "--color=yes",
    "--durations=0",
    "--strict-markers",
    "--doctest-modules",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]
log_cli = "True"
markers = [
    "slow: slow tests",
]
minversion = "6.0"
testpaths = "tests/"

[tool.coverage.report]
exclude_lines = [
    "pragma: nocover",
    "raise NotImplementedError",
    "raise NotImplementedError()",
    "if __name__ == .__main__.:",
]

# https://pypi.org/project/bumpver
[tool.bumpver]
current_version = "0.4.0"
version_pattern = "MAJOR.MINOR.PATCH[.PYTAGNUM]"
commit_message  = "Bump version {old_version} -> {new_version}"
commit          = true
tag             = false # no longer useful to tag here, must happen in create_publish_pr.yaml
push            = false

[tool.bumpver.file_patterns]
"pyproject.toml" = ['current_version = "{version}"', 'version = "{version}"']
"version.toml" = ['version = "{version}"']
"cyto_dl/__init__.py" = ['__version__ = "{version}"']
